+++ actix-http/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.12.0
+
 - Minimum supported Rust version (MSRV) is now 1.88.
 - Increase default HTTP/2 flow control window sizes. [#3638]
 - Expose configuration methods to improve upload throughput. [#3638]